Reimplement FileObjectPosix on top of the io package under Python2. Fixes #151.
