+++ greentest/test__all__.py
@@ -33,7 +33,9 @@ NOT_IMPLEMENTED = {
 COULD_BE_MISSING = {
     'socket': ['create_connection', 'RAND_add', 'RAND_egd', 'RAND_status']}
 
-NO_ALL = ['gevent.threading', 'gevent._util', 'gevent._socketcommon', 'gevent._fileobjectcommon',
+NO_ALL = ['gevent.threading', 'gevent._util',
+          'gevent._socketcommon',
+          'gevent._fileobjectcommon', 'gevent._fileobjectposix',
           'gevent._tblib']
